Add test run on review request
To avoid running slow/expesnive tests every time a change is pushed
to a draft PR we only run the test suite when a draft is ready for
review. However we also want to run the test suite when a PR has
been updated based on feedback and a review is re-requested.